move common parts of build config to separate file and merge with it
