+++ package.json
@@ -34,6 +34,7 @@
     "fake-indexeddb": "^3.1.2",
     "impunity": "^1.0.9",
     "mdn-polyfills": "^5.20.0",
+    "merge-options": "^3.0.4",
     "node-html-parser": "^4.0.0",
     "postcss-css-variables": "^0.18.0",
     "postcss-flexbugs-fixes": "^5.0.2",
